// apps/website/src/types/documentation.ts
export interface DocumentationClassMeta {
	line: number;
	file: string;
	path?: string;
	url?: string;
}

type DocumentationClassMethodMeta = DocumentationClassMeta;
type DocumentationClassMethodParameterMeta = DocumentationClassMeta;
type DocumentationClassMethodPropertyMeta = DocumentationClassMeta;
type DocumentationClassEventMeta = DocumentationClassMeta;
// ...
type DocumentationClassConstructorParameter = DocumentationParameter;
type DocumentationClassEventParameter = DocumentationParameter;

interface DocumentationClassMethodParameter extends DocumentationParameter {
	meta: DocumentationClassMethodParameterMeta;
}

interface DocumentationTypeDefinitionParameter extends DocumentationParameter {
	meta: DocumentationTypeDefinitionParameterMeta;
}

interface DocumentationProperty extends DocumentationParameter {}
type DocumentationTypeDefinitionProperty = DocumentationProperty;
type DocumentationClassPropertyProperty = DocumentationProperty;

export interface DocumentationClassProperty extends DocumentationProperty {
	props: DocumentationClassPropertyProperty[];
	meta: DocumentationClassMethodPropertyMeta;
}

interface DocumentationClassConstructor {
	name: string;
	params?: DocumentationClassConstructorParameter[];
}

export interface DocumentationClassMethod {
	name: string;
	description?: string;
	access?: string;
	async?: boolean;
	abstract?: boolean;
	deprecated?: boolean | string;
	inherited?: boolean;
	scope?: string;
	params?: DocumentationClassMethodParameter[];
	examples?: string[];
	returns?: unknown;
	returnsDescription?: string;
	throws?: string[];
	see?: string[];
	meta: DocumentationClassMethodMeta;
}

export interface DocumentationClassEvent {
	name: string;
	description: string;
	access?: string;
	deprecated?: boolean | string;
	params: DocumentationClassEventParameter[];
	see?: string[];
	meta: DocumentationClassEventMeta;
}

export interface DocumentationClass {
	name: string;
	deprecated?: string | boolean;
	description: string;
	construct: DocumentationClassConstructor;
	abstract: boolean;
	access?: string;
	extends: unknown;
	implements: unknown;
	props: DocumentationClassProperty[];
	methods: DocumentationClassMethod[];
	events: DocumentationClassEvent[];
	meta: DocumentationClassMeta;
}